+++ plugins/action/sda_fabric_zones_info.py
@@ -31,8 +31,8 @@ argument_spec.update(
     dict(
         id=dict(type="str"),
         siteId=dict(type="str"),
-        offset=dict(type="float"),
-        limit=dict(type="float"),
+        offset=dict(type="int"),
+        limit=dict(type="int"),
         headers=dict(type="dict"),
     )
 )
